fix-gdi-parsing take 2
Apparently, visual studio breaks posix compatibility and makes ssize_t unsigned of all things so it's better to never allow the len variable to go bellow zero even at the cost of a bit more complicated code. A alternative would be to cast to int, but i suppose that would give a spurious warning.