Fix CHD Loading on rpi/armv7+ due to -O3 optimisations - fixes #816
Force -O2 optimisation for any targets with arm dynarec due to crashes with -O3 and CHD files on arm.

Set LDFLAGS to CFLAGS for rpi otherwise the CPU flags are not included at link time with LTO. This avoids armv6 code
being generated on Raspbian on armv7 when LTO is used.